Added the neural network file and the LLM training file
This repo has the neural network file which handles all the neural network magic. You can choose which activation to use (RELU, tanh, or sigmoid), you can train the network with backpropergation, etc. It also has the vectoriser file which is responsible for vectorising all the words so that its not just one giant list of inputs for every word. You need to download the GenericsKB.tsv.gz and the GenericsKB-Best.tsv.gz from here: https://huggingface.co/datasets/generics_kb/tree/main/data. This is the training data. To be able to interact with this and talk to it, you need another network that has enough inputs for the prompt and what its prevously said, and enough outputs to correctly say what it needs to say. This is not optomised for speed much, but it should work on higher end computers. It is written completely from scratch so you dont need to have downloaded any external libraries except a base python install.